pyproject.toml
[project]
name = "algorithmic-trading-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alpaca-py>=0.39.0",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
]
